backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6